Split out graph related bits from elbo.js.
In order to make the file less unwieldly to work with.
